pyproject.toml
[project]
name = "ssl-wikichurches"
version = "0.1.0"
description = "Compare SSL model attention patterns against expert annotations on WikiChurches dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Existing
    "requests>=2.32.5",
    "tqdm>=4.67.2",
    # ML Core
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.56.0",
    "peft>=0.7.0",
    "timm>=1.0.20",
    # Data & IO
    "pillow>=10.0.0",
    "h5py>=3.10.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0.1",
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    # Data Exploration
    "polars>=1.0.0",
    "jupyterlab>=4.0.0",
    # Analysis
    "scikit-learn>=1.3.0",
    # Web API
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.lint.per-file-ignores]
"app/**/*.py" = ["E402"]  # Allow imports after sys.path modification

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
files = ["src", "app", "experiments"]

[[tool.mypy.overrides]]
module = ["h5py", "h5py.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "scipy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "scipy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sklearn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sklearn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torchvision"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torchvision.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tqdm", "tqdm.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ssl_attention.metrics.statistics",
    "ssl_attention.evaluation.linear_probe",
    "ssl_attention.models.resnet50",
    "ssl_attention.evaluation.fine_tuning",
]
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssl_attention"]
